requirements.txt
streamlit==0.56.0
pandas
numpy
scipy==1.3.1
plotly==4.5.4
tqdm
pyyaml
pytest
ipykernel
requests